Added automapper profiles

+++ TravelHelper.BusinessLayer/HotelManagement/Commands/UpdateHotelCommand.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
+using BusinessLayer.HotelManagement.DTO;
 using BusinessLayer.Shared;
-using BusinessLayer.Shared.DTO;
 using MediatR;
 
 namespace BusinessLayer.HotelManagement.Commands
@@ -13,9 +13,9 @@ namespace BusinessLayer.HotelManagement.Commands
         public string Country { get; set; }
         public string City { get; set; }
         public string Address { get; set; }
-        public decimal PricePerDay { get; set; }
-        public decimal Longitude { get; set; }
-        public decimal Latitude { get; set; }
+        public double PricePerDay { get; set; }
+        public double Longitude { get; set; }
+        public double Latitude { get; set; }
         public ICollection<ImageDto> Images { get; set; }
     }
 }
